fin-us/scripts/templates/requirements_integrated.txt
# NeMo integrate build: MCP + FastAPI + NAT proxy (no reference OpenAI/Anthropic orchestration).
fastapi
uvicorn[standard]
pydantic
python-dotenv
mcp
pyyaml
httpx
